The project was based on the image python: 3.10-alpine
An issue has been found related to the lack of a GNU version (glibc) of
the C standard library, which is in fact required by every program
written in C, including Python.
But Alpine uses `musl`, and since those binaries are for `glibc`,
they simply don't work.
To solve the problem, I had to execute the script: "apk add build-base".

Link to the description of the basis of the problem:
https://habr.com/ru/post/486202/
Link to found solution:
https://github.com/MagicStack/uvloop/issues/168#issuecomment-581012557